+++ frontend/src/components/recipes/search_results_index_container.js
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 
 import SearchResultsIndex from './search_results_index';
-import { fetchRecipes } from '../../actions/recipe_actions';
+import { fetchRecipes } from '../../actions/recipes_actions';
 import { asArray } from '../../reducers/selectors';
 
 const mapStateToProps = state => ({
