+++ frontend/src/reducer/entities_reducer.js
